test: made test script easier to digest & reuse for feature deployments
